Stop event from firing multiple times for the same scenario
A card can collide with the area more than once; many of the campaign memory bags caused a double collision on placement, and it would happen if the user picked up or flipped the card.

This would cause any behavior that triggered from the event to each time, so suppress further events which don't actually change the scenario.