Handle duplicate connections
Some icon/ability combinations on locations results in duplicate connections.  The connection logic was parsing this as both a bidirectional and one-way connection, resulting in a one-way being drawn when it shouldn't be.

This adds a simple check to recommit an existing bidirectional connection, rather than drawing a new one-way for it.