First Pass at Deck Builder
Got the initial functionality of the deck builder working, is now capable of presenting faction-specific cards of the correct levels to the player.

Also created the index to be able to more easily retrieve cards by trait.

Still have to edit card placement and spwaning to be more tidy, and move off to the middle section since there are too many card options at level 0.